+++ src/AbstractNotification.php
@@ -11,42 +11,9 @@
 
 namespace Notify;
 
-use Notify\Strategy\StrategyInterface;
-use Notify\Exception\NotificationStrategyNotSuppliedException;
-
 /**
  * @author Nikola Posa <posa.nikola@gmail.com>
  */
 abstract class AbstractNotification implements NotificationInterface
 {
-    /**
-     * @var StrategyInterface
-     */
-    private static $defaultStrategy = null;
-
-    /**
-     * @param StrategyInterface $defaultStrategy
-     */
-    public static function setDefaultStrategy(StrategyInterface $defaultStrategy)
-    {
-        self::$defaultStrategy = $defaultStrategy;
-    }
-
-    public static function resetDefaultStrategy()
-    {
-        self::$defaultStrategy = null;
-    }
-
-    public function __invoke(StrategyInterface $strategy = null)
-    {
-        if (null === $strategy) {
-            if (null === self::$defaultStrategy) {
-                throw NotificationStrategyNotSuppliedException::forNotification($this);
-            }
-
-            $strategy = self::$defaultStrategy;
-        }
-
-        $strategy->handle($this);
-    }
 }