+++ CHANGELOG.md
@@ -4,4 +4,5 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Backwards-incompatible changes
+- Removed `NotificationInterface::__invoke()` method
 - `MessageSender` in favor of `SendService` naming for message sender implementations.